dummy out block to include error
knitr doesn't show the error for this block, even with error=TRUE, which is bad because the point of the block is to show the error. I'm not sure why knitr doesn't show this error; this is the quick hack solution to just fake the block so that it looks the way it does when trying it at the console.